+++ plugins/catalog-backend/src/processing/DefaultCatalogProcessingEngine.test.ts
@@ -70,6 +70,8 @@ describe('DefaultCatalogProcessingEngine', () => {
       orchestrator: orchestrator,
       stitcher: stitcher,
       createHash: () => hash,
+      scheduler: mockServices.scheduler(),
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -136,7 +138,9 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -219,7 +223,9 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -296,7 +302,9 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -355,8 +363,10 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
       pollingIntervalMs: 100,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -470,8 +480,10 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
       pollingIntervalMs: 100,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -575,8 +587,10 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
       pollingIntervalMs: 100,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -658,8 +672,10 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
       pollingIntervalMs: 100,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));
@@ -746,8 +762,10 @@ describe('DefaultCatalogProcessingEngine', () => {
       knex: {} as any,
       orchestrator: orchestrator,
       stitcher: stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => hash,
       pollingIntervalMs: 100,
+      events: mockServices.events.mock(),
     });
 
     db.transaction.mockImplementation(cb => cb((() => {}) as any));