+++ plugins/catalog-backend/src/tests/integration.test.ts
@@ -244,6 +244,7 @@ class TestHarness {
     const processingDatabase = new DefaultProcessingDatabase({
       database: options.db,
       logger,
+      events: mockServices.events.mock(),
       refreshInterval: () => 0.05,
     });
 
@@ -273,7 +274,6 @@ class TestHarness {
       logger,
       parser: defaultEntityDataParser,
       policy: EntityPolicies.allOf([]),
-      legacySingleProcessorValidation: false,
     });
     const stitcher = DefaultStitcher.fromConfig(config, {
       knex: options.db,
@@ -296,12 +296,14 @@ class TestHarness {
       knex: options.db,
       orchestrator,
       stitcher,
+      scheduler: mockServices.scheduler(),
       createHash: () => createHash('sha1'),
       pollingIntervalMs: 50,
       onProcessingError: event => {
         proxyProgressTracker.reportError(event.unprocessedEntity, event.errors);
       },
       tracker: proxyProgressTracker,
+      events: mockServices.events.mock(),
     });
 
     const refresh = new DefaultRefreshService({ database: catalogDatabase });