+++ plugins/scaffolder-backend/src/scaffolder/tasks/DatabaseTaskStore.ts
@@ -85,8 +85,6 @@ export type RawDbTaskEventRow = {
 
 /**
  * DatabaseTaskStore
- * @deprecated this type is deprecated, and there will be a new way to create Workers in the next major version.
- * @public
  */
 export type DatabaseTaskStoreOptions = {
   database: DatabaseService | Knex;
@@ -118,8 +116,6 @@ const parseSqlDateToIsoString = <T>(input: T): T | string => {
 
 /**
  * DatabaseTaskStore
- * @deprecated this type is deprecated, and there will be a new way to create Workers in the next major version.
- * @public
  */
 export class DatabaseTaskStore implements TaskStore {
   private readonly db: Knex;
@@ -723,7 +719,7 @@ export class DatabaseTaskStore implements TaskStore {
     });
   }
 
-  async retryTask?(options: {
+  async retryTask(options: {
     secrets?: TaskSecrets;
     taskId: string;
   }): Promise<void> {