+++ plugins/scaffolder/src/alpha/fields/RepoUrlPicker.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Backstage Authors
+ * Copyright 2025 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 import { createFormField } from '@backstage/plugin-scaffolder-react/alpha';
-import { RepoUrlPicker as Component } from '../../components/fields/RepoUrlPicker/RepoUrlPicker';
 import {
-  RepoUrlPickerFieldSchema,
   repoPickerValidation,
-} from '../../components';
+  RepoUrlPicker as Component,
+  RepoUrlPickerFieldSchema,
+} from '../../components/fields/RepoUrlPicker';
 
 export const RepoUrlPicker = createFormField({
   component: Component,