+++ plugins/scaffolder/src/components/fields/RepoUrlPicker/validation.ts
@@ -22,7 +22,6 @@ import { scmIntegrationsApiRef } from '@backstage/integration-react';
  * The validation function for the `repoUrl` that is returned from the
  * field extension. Ensures that you have all the required fields filled for
  * the different providers that exist.
- *
  * @public
  */
 export const repoPickerValidation = (