+++ packages/app-next-example-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # app-next-example-plugin
 
+## 0.0.28-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.18.2-next.0
+  - @backstage/frontend-plugin-api@0.12.1-next.0
+
 ## 0.0.27
 
 ### Patch Changes