+++ packages/frontend-app-api/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage/frontend-app-api
 
+## 0.13.1-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.4-next.0
+  - @backstage/core-app-api@1.19.1-next.0
+  - @backstage/core-plugin-api@1.11.1-next.0
+  - @backstage/frontend-defaults@0.3.2-next.1
+  - @backstage/frontend-plugin-api@0.12.1-next.1
+
+## 0.13.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-defaults@0.3.2-next.0
+  - @backstage/frontend-plugin-api@0.12.1-next.0
+  - @backstage/config@1.3.3
+  - @backstage/core-app-api@1.19.0
+  - @backstage/core-plugin-api@1.11.0
+  - @backstage/errors@1.2.7
+  - @backstage/types@1.2.2
+  - @backstage/version-bridge@1.0.11
+
 ## 0.13.0
 
 ### Minor Changes