+++ plugins/auth-backend-module-cloudflare-access-provider/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-auth-backend-module-cloudflare-access-provider
 
+## 0.4.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.4-next.0
+  - @backstage/backend-plugin-api@1.4.4-next.0
+  - @backstage/plugin-auth-node@0.6.8-next.0
+
 ## 0.4.7
 
 ### Patch Changes