+++ plugins/auth-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-auth-backend
 
+## 0.25.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.4-next.0
+  - @backstage/backend-plugin-api@1.4.4-next.0
+  - @backstage/plugin-auth-node@0.6.8-next.0
+  - @backstage/plugin-catalog-node@1.19.1-next.0
+
 ## 0.25.4
 
 ### Patch Changes