+++ plugins/notifications-backend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-notifications-backend
 
+## 0.5.11-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.3.4-next.0
+  - @backstage/backend-plugin-api@1.4.4-next.0
+  - @backstage/plugin-auth-node@0.6.8-next.0
+  - @backstage/plugin-notifications-common@0.1.1-next.0
+  - @backstage/plugin-signals-node@0.1.25-next.0
+  - @backstage/plugin-catalog-node@1.19.1-next.0
+  - @backstage/plugin-events-node@0.4.16-next.0
+  - @backstage/plugin-notifications-node@0.2.20-next.0
+
 ## 0.5.10
 
 ### Patch Changes