+++ plugins/scaffolder-common/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-scaffolder-common
 
+## 1.7.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration@1.18.1-next.1
+  - @backstage/plugin-permission-common@0.9.2-next.0
+
+## 1.7.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration@1.18.1-next.0
+  - @backstage/catalog-model@1.7.5
+  - @backstage/errors@1.2.7
+  - @backstage/types@1.2.2
+  - @backstage/plugin-permission-common@0.9.1
+
 ## 1.7.1
 
 ### Patch Changes