+++ plugins/scaffolder-node-test-utils/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-scaffolder-node-test-utils
 
+## 0.3.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.4.4-next.0
+  - @backstage/backend-test-utils@1.9.1-next.1
+  - @backstage/plugin-scaffolder-node@0.12.0-next.1
+
+## 0.3.4-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-node@0.12.0-next.0
+  - @backstage/backend-test-utils@1.9.1-next.0
+  - @backstage/backend-plugin-api@1.4.3
+  - @backstage/types@1.2.2
+
 ## 0.3.3
 
 ### Patch Changes